changed: put tag videodatabase methods in separate file
