+++ xbmc/video/database/VideoDatabase.cpp
@@ -1368,14 +1368,6 @@ int CVideoDatabase::AddUniqueIDs(int mediaId, const char *mediaType, const CVide
   return uniqueid;
 }
 
-int CVideoDatabase::AddTag(const std::string& name)
-{
-  if (name.empty())
-    return -1;
-
-  return AddToTable("tag", "tag_id", "name", name);
-}
-
 int CVideoDatabase::AddActor(const std::string& name, const std::string& thumbURLs, const std::string &thumb)
 {
   try
@@ -1494,31 +1486,6 @@ void CVideoDatabase::UpdateActorLinksToItem(int mediaId, const std::string& medi
   m_pDS->exec(sql);
 
   AddActorLinksToItem(mediaId, mediaType, field, values);
-}
-
-//****Tags****
-void CVideoDatabase::AddTagToItem(int media_id, int tag_id, const std::string &type)
-{
-  if (type.empty())
-    return;
-
-  AddToLinkTable(media_id, type, "tag", tag_id);
-}
-
-void CVideoDatabase::RemoveTagFromItem(int media_id, int tag_id, const std::string &type)
-{
-  if (type.empty())
-    return;
-
-  RemoveFromLinkTable(media_id, type, "tag", tag_id);
-}
-
-void CVideoDatabase::RemoveTagsFromItem(int media_id, const std::string &type)
-{
-  if (type.empty())
-    return;
-
-  m_pDS2->exec(PrepareSQL("DELETE FROM tag_link WHERE media_id=%d AND media_type='%s'", media_id, type.c_str()));
 }
 
 //****Actors****
@@ -2493,32 +2460,6 @@ void CVideoDatabase::DeleteStreamDetails(int idFile)
   m_pDS->exec(PrepareSQL("DELETE FROM streamdetails WHERE idFile = %i", idFile));
 }
 
-void CVideoDatabase::DeleteTag(int idTag, VIDEODB_CONTENT_TYPE mediaType)
-{
-  try
-  {
-    if (m_pDB.get() == NULL || m_pDS.get() == NULL)
-      return;
-
-    std::string type;
-    if (mediaType == VIDEODB_CONTENT_MOVIES)
-      type = MediaTypeMovie;
-    else if (mediaType == VIDEODB_CONTENT_TVSHOWS)
-      type = MediaTypeTvShow;
-    else if (mediaType == VIDEODB_CONTENT_MUSICVIDEOS)
-      type = MediaTypeMusicVideo;
-    else
-      return;
-
-    std::string strSQL = PrepareSQL("DELETE FROM tag_link WHERE tag_id = %i AND media_type = '%s'", idTag, type.c_str());
-    m_pDS->exec(strSQL);
-  }
-  catch (...)
-  {
-    CLog::Log(LOGERROR, "%s (%i) failed", __FUNCTION__, idTag);
-  }
-}
-
 void CVideoDatabase::GetDetailsFromDB(std::unique_ptr<Dataset> &pDS, int min, int max, const SDbTableOffsets *offsets, CVideoInfoTag &details, int idxOffset)
 {
   GetDetailsFromDB(pDS->get_sql_record(), min, max, offsets, details, idxOffset);
@@ -2892,28 +2833,6 @@ void CVideoDatabase::GetCast(int media_id, const std::string &media_type, std::v
         info.thumb = m_pDS2->fv(4).get_asString();
         cast.emplace_back(std::move(info));
       }
-      m_pDS2->next();
-    }
-    m_pDS2->close();
-  }
-  catch (...)
-  {
-    CLog::Log(LOGERROR, "%s(%i,%s) failed", __FUNCTION__, media_id, media_type.c_str());
-  }
-}
-
-void CVideoDatabase::GetTags(int media_id, const std::string &media_type, std::vector<std::string> &tags)
-{
-  try
-  {
-    if (!m_pDB.get()) return;
-    if (!m_pDS2.get()) return;
-
-    std::string sql = PrepareSQL("SELECT tag.name FROM tag INNER JOIN tag_link ON tag_link.tag_id = tag.tag_id WHERE tag_link.media_id = %i AND tag_link.media_type = '%s' ORDER BY tag.tag_id", media_id, media_type.c_str());
-    m_pDS2->query(sql);
-    while (!m_pDS2->eof())
-    {
-      tags.emplace_back(m_pDS2->fv(0).get_asString());
       m_pDS2->next();
     }
     m_pDS2->close();
@@ -4468,11 +4387,6 @@ bool CVideoDatabase::GetNavCommon(const std::string& strBaseDir, CFileItemList& 
   return false;
 }
 
-bool CVideoDatabase::GetTagsNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
-{
-  return GetNavCommon(strBaseDir, items, "tag", idContent, filter, countOnly);
-}
-
 bool CVideoDatabase::GetWritersNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
 {
   return GetPeopleNav(strBaseDir, items, "writer", idContent, filter, countOnly);
@@ -5467,11 +5381,6 @@ std::string CVideoDatabase::GetGenreById(int id)
 std::string CVideoDatabase::GetCountryById(int id)
 {
   return GetSingleValue("country", "name", PrepareSQL("country_id=%i", id));
-}
-
-std::string CVideoDatabase::GetTagById(int id)
-{
-  return GetSingleValue("tag", "name", PrepareSQL("tag_id = %i", id));
 }
 
 std::string CVideoDatabase::GetPersonById(int id)
